PC source code in only one file for all OS - Windows, Linux and Mac. Included delay in Linux for slow down. The BlkTot not multiple of 512 implemented for all OS. It was tested in Windows and Linux.
